requirements.txt.txt
streamlit
opencv-python-headless
numpy
pillow
torch
torchvision
ultralytics
clip-anytorch
pandas
scikit-image
plotly
streamlit-image-comparison
